refactor(extensions): Move string validation extensions
Relocates `NotNullOrEmpty` and `NotNullOrWhitespace` from `CommonExtensions` to `StringExtensions` for better type-specific organization.
