+++ src/GitVersion.Core/Extensions/StringExtensions.cs
@@ -1,5 +1,6 @@
 using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
+using System.Runtime.CompilerServices;
 
 namespace GitVersion.Extensions;
 
@@ -38,6 +39,31 @@ public static class StringExtensions
 
         /// <inheritdoc cref="string.IsNullOrWhiteSpace"/>
         public bool IsNullOrWhiteSpace() => string.IsNullOrWhiteSpace(value);
+    }
+
+    extension([NotNull] string? value)
+    {
+        /// <summary>Throws <see cref="ArgumentException"/> when <paramref name="value"/> is <see langword="null"/> or empty; otherwise returns the value.</summary>
+        public string NotNullOrEmpty([CallerArgumentExpression(nameof(value))] string name = "")
+        {
+            if (string.IsNullOrEmpty(value))
+            {
+                throw new ArgumentException("The parameter is null or empty.", name);
+            }
+
+            return value;
+        }
+
+        /// <summary>Throws <see cref="ArgumentException"/> when <paramref name="value"/> is <see langword="null"/>, empty, or whitespace; otherwise returns the value.</summary>
+        public string NotNullOrWhitespace([CallerArgumentExpression(nameof(value))] string name = "")
+        {
+            if (string.IsNullOrWhiteSpace(value))
+            {
+                throw new ArgumentException("The parameter is null or empty or contains only white space.", name);
+            }
+
+            return value;
+        }
     }
 
     internal static string ToPascalCase(this TextInfo textInfo, string input)
